--- src/routes/faq.js
@@ -1,57 +1,0 @@
-export async function GET() {
-    return {
-        status: 200,
-        body: {
-            faqs: [
-                {
-                    "q": "How much does this site cost?",
-                    "a": "Using this site is COMPLETELY FREE, and it will always stay this way. I pay for all of the costs out of my own pocket. That being said, please donate if you want to help out with the costs of this site."
-                },
-                {
-                    "q": "Why should I trust you?",
-                    "a": "I am the most experienced person when it comes to recommending Gambit plays. I’ve been doing this since February 2020. But, don’t take my word for it. Many of my followers use the same method as I do and can report huge earnings as well!"
-                },
-                {
-                    "q": "What is the difference between reward and odds?",
-                    "a": "Odds are the raw payout from the sportsbook that provides Gambit's data (Intertops). Reward is what Gambit is willing to pay you, the user, for a line. Reward will be lower than odds in most cases because Gambit is interested in making a profit from their operation. It's kind of like a 'vig' except it's charged if you win (not if you lose)."
-                },
-                {
-                    "q": "What is Gambit?",
-                    "a": "Gambit Rewards is a sports “betting” website, where you can bet on sports/eSports games using a special currency called “Gambit tokens”. It’s really only worth using if you follow the strategies on this website, as the cuts they take make it quite poor for actual sports betting. If you want to actually bet on sports, there are many other (better) websites out there."
-                },
-                {
-                    "q": "How does Gambit work?",
-                    "a": "The entire premise of the profitability of using Gambit is based on the fact that you can get 300 Gambit tokens (i.e. $3.00 USD) for less than 300 Swagbucks (i.e. < $3.00 USD) – as of 11/01/2020, 1000 Gambit tokens = 900 SB. As long as you make more than 900 tokens from any bet of 1000 tokens, that bet is profitable. Plus, you can choose the amount of Gambit tokens you convert back to Swagbucks at one time. This is super useful for getting the Swagbucks daily goals, the rewards for which can be up to $200/month for doing nothing!"
-                },
-                {
-                    "q": "Where do I sign up for Gambit?",
-                    "a": "You can sign up for Gambit <a href='https://go.unixfy.net/gambit' rel='noopener'>here</a> (referral link, you will get $1.00 free upon signing up) or <a href='http://app.gambitrewards.com/' rel='noopener'>here</a> (non-referral link). Please note: your Gambit account email must be the same as your Swagbucks account email to receive tokens."
-                },
-                {
-                    "q": "How often is this site updated?",
-                    "a": "This website is updated automatically, within 5 minutes of GambitRewards’s website being updated with new bets."
-                },
-                {
-                    "q": "Can you tell me about the different betting methods?",
-                    "a": "Risky betting: You bet on only one team (or side) of each game. This team is picked specifically by me because it has a relatively low chance of losing (usually less than 10% chance of losing). This will yield, by far, the most rewards, but also carries some degree of risk. Ultimately, this method can still be profitable even if you lose some bets. <br><br>Med-risk betting: You bet on 2 of 3 teams in a soccer match. The two teams you bet on are picked specifically because they have a relatively high chance of winning. This will yield more rewards than no-risk betting but less rewards than risky betting. <br><br>No-risk betting: You bet on every team (or side) of each game. The amount you bet on each team is calculated such that you will receive the same amount of tokens, regardless of which team wins. This totally eliminates the risk, but also means lower rewards (about half to 2/3 of the rewards of risky betting). <br><br>Overall, the reward you earn has a strong correlation with the amount of risk in a bet."
-                },
-                {
-                    "q": "Why is the 'play' button greyed out?",
-                    "a": "You need to bet at least 10 tokens on any side in order to submit your bet."
-                },
-                {
-                    "q": "Why can’t I bet on multiple teams at one time?",
-                    "a": "This is just a limitation of the GambitRewards website \uD83D\uDE41. <br>Workaround: You need to type in a token amount for one team, click “Play now”, and repeat it for every team you wish to bet on. Inconvenient!"
-                },
-                {
-                    "q": "I accidentally created a Gambit account with a different email than my Swagbucks account. What do I do?",
-                    "a": "You can either change the email address on your Swagbucks account, or just sign up for a new Gambit account with the correct email address."
-                },
-                {
-                    "q": "What are these “Instant win” games on Gambit?",
-                    "a": "They’re similar to online casino games. Like those games, you’ll never profit from them, so just avoid them."
-                }
-            ]
-        }
-    }
-}
